requirements.txt
bottle
lxml
Pillow
protobuf
pyunpack
reportlab
Requests
google-cloud-vision
patool
